NeighborlyAPI/getAdvertisements/__init__.py
import azure.functions as func
import pymongo
import json
import sys
from bson.json_util import dumps
import os

def main(req: func.HttpRequest) -> func.HttpResponse:

    try:
        url = os.environ['MyDbConnection']
        client = pymongo.MongoClient(url)
        database = client['project3db']
        collection = database['advertisements']


        result = collection.find({})
        result = dumps(result)

        return func.HttpResponse(result, mimetype="application/json", charset='utf-8')
    except:
        print(sys.exc_info())
        print("could not connect to mongodb")
        return func.HttpResponse("could not connect to mongodb",
                                 status_code=400)

