NeighborlyAPI/requirements.txt
azure-functions
pymongo
dnspython
